+++ BlazorWasm/BinaryTool/BinaryTool.Blazor/Pages/ParseBinary.razor
@@ -1,4 +1,5 @@
-@using BinaryTool.Binary;
+@using BinaryTool.Binary.Parser;
+@using BinaryTool.Binary.Formatter;
 
 <Selector Label="Parser" Items="IParser.DefaultParsers" @bind-SelectedItem="SelectedParser" />
