+++ BlazorWasm/BinaryTool/BinaryTool/Binary/Formatter/NumberFormat.cs
@@ -1,0 +1,9 @@
+﻿namespace BinaryTool.Binary.Formatter;
+
+public enum NumberFormat
+{
+    Decimal,
+    Hex,
+    Hex0x,
+    Hex0X,
+}
