+++ BlazorWasm/BinaryTool/BinaryTool/Binary/Parser/DecimalParser.cs
@@ -1,0 +1,9 @@
+﻿namespace BinaryTool.Binary.Parser;
+
+public class DecimalParser : IParser
+{
+    public static readonly DecimalParser Instance = new();
+
+    public string Description => "decimal numbers";
+    public byte[] Parse(string input) => ParserHelper.Parse(input, false);
+}
