+++ BlazorWasm/BinaryTool/BinaryTool/Binary/Parser/HexParser.cs
@@ -1,0 +1,9 @@
+﻿namespace BinaryTool.Binary.Parser;
+
+public class HexParser : IParser
+{
+    public static readonly HexParser Instance = new();
+
+    public string Description => "hexadecimal numbers";
+    public byte[] Parse(string input) => ParserHelper.Parse(input, true);
+}
