+++ BlazorWasm/BinaryTool/BinaryTool/Dom/ILoader.cs
@@ -6,11 +6,11 @@ public interface ILoader : IDescriptiveItem
 {
     (List<DomSpan> spans, bool isBinary) Parse(byte[] data);
 
-    public static readonly ImmutableArray<ILoader> DefaultLoaders = ImmutableArray.Create(new ILoader[]
+    public static readonly ImmutableArray<ILoader> Loaders = ImmutableArray.Create(new ILoader[]
     {
         JsonLoader.Instance,
         MessagePackLoader.Instance,
     });
 
-    public static readonly ImmutableArray<ILoader> DefaultLoadersWithAuto = ImmutableArray.CreateRange(DefaultLoaders.Prepend(AutoDetectLoader.Instance));
+    public static readonly ImmutableArray<ILoader> DefaultLoaders = ImmutableArray.CreateRange(Loaders.Prepend(AutoDetectLoader.Instance));
 }
